new meta.function_definition table, for use in bundling postgresql functions
this is working around a bug in meta.function where argument names (which the function code references) are not included in the parameters field.  they're included in the meta.function_parameters view, which is also just a crappy hack.  generally, meta.function needs to be rewritten and then meta.function_definition and meta.function_parameter can be dropped as well. however, this would have potentially huge ramifications throughout endpoint.select_rows_function, which also needs a rewrite.  and we're just not taking that on right now.  ok.